+++ revdep/checks.noindex/detrendr/old/detrendr.Rcheck/tests/testthat/test-dir-detrend.R
@@ -1,14 +1,15 @@
 test_that("detrending entire derectories works", {
   skip_if(getRversion() < "3.6.0")
   skip_on_cran()
-  cwd <- setwd(tempdir())
+  cwd <- setwd(tempdir(check = TRUE))
   on.exit(setwd(cwd))
   orig_files <- c(
-    system.file("img", "2ch_ij.tif", package = "ijtiff"),
+    system.file("extdata", "2ch_ij.tif", package = "detrendr"),
     system.file("extdata", "bleached.tif", package = "detrendr")
   )
   file.copy(orig_files, ".")
-  orig_imgs <- purrr::map(orig_files, ijtiff::read_tif, msg = FALSE)
+  orig_imgs <- purrr::map(orig_files, ijtiff::read_tif, msg = FALSE) %>%
+    magrittr::set_names(basename(orig_files))
   detrendeds <- try(stop("eee"), silent = TRUE)
   set.seed(1)
   while (class(detrendeds) == "try-error") {
@@ -57,8 +58,13 @@ test_that("detrending entire derectories works", {
   }
   detrendeds_dir <- dir(pattern = "detrended.*tif", recursive = TRUE) %>%
     purrr::map(ijtiff::read_tif, msg = FALSE)
-  expect_equal(purrr::map(detrendeds, dim), purrr::map(detrendeds_dir, dim))
-  expect_equal(unlist(detrendeds), unlist(detrendeds_dir))
+  expect_equivalent(
+    purrr::map(detrendeds, dim),
+    purrr::map(detrendeds_dir, dim)
+  )
+  expect_equivalent(unlist(detrendeds), unlist(detrendeds_dir),
+    tolerance = 1
+  )
   if (get_os() == "mac") {
     expect_equal(
       dir("detrended"),
@@ -112,8 +118,11 @@ test_that("detrending entire derectories works", {
   )
   detrendeds_dir <- dir(pattern = "detrended.*tif", recursive = TRUE) %>%
     purrr::map(ijtiff::read_tif, msg = FALSE)
-  expect_equal(purrr::map(detrendeds, dim), purrr::map(detrendeds_dir, dim))
-  expect_equal(unlist(detrendeds), unlist(detrendeds_dir))
+  expect_equivalent(
+    purrr::map(detrendeds, dim),
+    purrr::map(detrendeds_dir, dim)
+  )
+  expect_equivalent(unlist(detrendeds), unlist(detrendeds_dir), tolerance = 1)
   if (get_os() == "mac") {
     expect_equal(
       dir("detrended"),
@@ -151,6 +160,54 @@ test_that("detrending entire derectories works", {
   detrendeds <- purrr::map(orig_imgs, autothresholdr::mean_stack_thresh,
     method = "tri"
   ) %>%
+    purrr::map(img_detrend_rh, swaps = 222)
+  set.seed(1)
+  dir_detrend_rh(swaps = 222, thresh = "tri", msg = FALSE)
+  detrendeds_dir <- dir(pattern = "detrended.*tif$", recursive = TRUE) %>%
+    purrr::map(ijtiff::read_tif, msg = FALSE)
+  expect_equivalent(
+    purrr::map(detrendeds, dim),
+    purrr::map(detrendeds_dir, dim)
+  )
+  expect_equivalent(unlist(detrendeds), unlist(detrendeds_dir))
+  if (get_os() == "mac") {
+    expect_equal(
+      dir("detrended"),
+      paste0(
+        c(
+          "2ch_ij_detrended_thresh=Triangle=0.6,Triangle=0.6_",
+          "bleached_detrended_thresh=Triangle=41.622_"
+        ),
+        c(
+          "robinhood_swaps=222,222.tif",
+          "robinhood_swaps=222.tif"
+        )
+      )
+    )
+  }
+  if (get_os() == "linux") {
+    expect_equal(
+      dir("detrended"),
+      paste0(
+        c(
+          "2ch_ij_detrended_thresh=Triangle=0.6,Triangle=0.6_",
+          "bleached_detrended_thresh=Triangle=41.622_"
+        ),
+        c(
+          "robinhood_swaps=222,222.tif",
+          "robinhood_swaps=222.tif"
+        )
+      )
+    )
+  }
+  filesstrings::dir.remove("detrended")
+  set.seed(1)
+  file.remove("2ch_ij.tif")
+  detrendeds <- purrr::map(
+    orig_imgs[list.files(pattern = ".tif$")],
+    autothresholdr::mean_stack_thresh,
+    method = "tri"
+  ) %>%
     purrr::map(~ suppressWarnings(img_detrend_polynom(.,
       degree = 2,
       purpose = "ff"
@@ -162,18 +219,19 @@ test_that("detrending entire derectories works", {
   ))
   detrendeds_dir <- dir(pattern = "detrended.*tif", recursive = TRUE) %>%
     purrr::map(ijtiff::read_tif, msg = FALSE)
-  expect_equal(purrr::map(detrendeds, dim), purrr::map(detrendeds_dir, dim))
-  expect_equal(unlist(detrendeds), unlist(detrendeds_dir))
+  expect_equivalent(
+    purrr::map(detrendeds, dim),
+    purrr::map(detrendeds_dir, dim)
+  )
+  expect_equivalent(unlist(detrendeds), unlist(detrendeds_dir), tolerance = 1)
   if (get_os() == "mac") {
     expect_equal(
       dir("detrended"),
       paste0(
         c(
-          "2ch_ij_detrended_thresh=Triangle=0.6,Triangle=0.6_",
           "bleached_detrended_thresh=Triangle=41.622_"
         ),
         c(
-          "polynomial_for_FFS_degree=2,2.tif",
           "polynomial_for_FFS_degree=2.tif"
         )
       )
@@ -184,54 +242,10 @@ test_that("detrending entire derectories works", {
       dir("detrended"),
       paste0(
         c(
-          "2ch_ij_detrended_thresh=Triangle=0.6,Triangle=0.6_",
           "bleached_detrended_thresh=Triangle=41.622_"
         ),
         c(
-          "polynomial_for_FFS_degree=2,2.tif",
           "polynomial_for_FFS_degree=2.tif"
-        )
-      )
-    )
-  }
-  filesstrings::dir.remove("detrended")
-  set.seed(1)
-  detrendeds <- purrr::map(orig_imgs, autothresholdr::mean_stack_thresh,
-    method = "tri"
-  ) %>%
-    purrr::map(img_detrend_rh, swaps = 222)
-  set.seed(1)
-  dir_detrend_rh(swaps = 222, thresh = "tri", msg = FALSE)
-  detrendeds_dir <- dir(pattern = "detrended.*tif$", recursive = TRUE) %>%
-    purrr::map(ijtiff::read_tif, msg = FALSE)
-  expect_equal(purrr::map(detrendeds, dim), purrr::map(detrendeds_dir, dim))
-  expect_equal(unlist(detrendeds), unlist(detrendeds_dir))
-  if (get_os() == "mac") {
-    expect_equal(
-      dir("detrended"),
-      paste0(
-        c(
-          "2ch_ij_detrended_thresh=Triangle=0.6,Triangle=0.6_",
-          "bleached_detrended_thresh=Triangle=41.622_"
-        ),
-        c(
-          "robinhood_swaps=222,222.tif",
-          "robinhood_swaps=222.tif"
-        )
-      )
-    )
-  }
-  if (get_os() == "linux") {
-    expect_equal(
-      dir("detrended"),
-      paste0(
-        c(
-          "2ch_ij_detrended_thresh=Triangle=0.6,Triangle=0.6_",
-          "bleached_detrended_thresh=Triangle=41.622_"
-        ),
-        c(
-          "robinhood_swaps=222,222.tif",
-          "robinhood_swaps=222.tif"
         )
       )
     )
@@ -242,6 +256,7 @@ test_that("detrending entire derectories works", {
 })
 
 test_that("file_detrend() deals with other directories correctly", {
+  skip_on_cran()
   setwd(tempdir())
   filesstrings::create_dir("tempwithintemp")
   file.copy(