+++ revdep/library.noindex/ijtiff/old/ijtiff/NEWS.md
@@ -1,3 +1,15 @@
+# `ijtiff` 2.1.2
+
+## BUG FIXES
+* Fix some typos in the vignettes.
+
+
+# `ijtiff` 2.1.1
+
+## BUG FIXES
+* Fix a PROTECTion error.
+
+
 # `ijtiff` 2.1.0
 
 ## NEW FEATURES